requirements.txt
flask
azure-cosmosdb-table
azure-storage-blob
stripe
tensorflow
sklearn
pandas
numpy
opencv-python